+++ pkg/registry/apis/provisioning/request_test.go
@@ -1,11 +1,9 @@
 package provisioning
 
 import (
-	"net/http"
 	"net/http/httptest"
 	"strings"
 	"testing"
-	"time"
 )
 
 func TestReadBody(t *testing.T) {
@@ -94,46 +92,6 @@ func TestIsJSONContentType(t *testing.T) {
 			r.Header.Set("Content-Type", tt.contentType)
 			if got := isJSONContentType(r); got != tt.want {
 				t.Errorf("isJSONContentType() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestWithTimeout(t *testing.T) {
-	handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		select {
-		case <-r.Context().Done():
-			w.WriteHeader(http.StatusGatewayTimeout)
-		case <-time.After(50 * time.Millisecond):
-			w.WriteHeader(http.StatusOK)
-		}
-	})
-
-	tests := []struct {
-		name       string
-		timeout    time.Duration
-		wantStatus int
-	}{
-		{
-			name:       "request completes",
-			timeout:    100 * time.Millisecond,
-			wantStatus: http.StatusOK,
-		},
-		{
-			name:       "request times out",
-			timeout:    10 * time.Millisecond,
-			wantStatus: http.StatusGatewayTimeout,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			w := httptest.NewRecorder()
-			r := httptest.NewRequest("GET", "/", nil)
-			withTimeout(handler, tt.timeout).ServeHTTP(w, r)
-
-			if w.Code != tt.wantStatus {
-				t.Errorf("withTimeout() status = %v, want %v", w.Code, tt.wantStatus)
 			}
 		})
 	}
